// src/api/orval/model/overviewGrouping.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export type OverviewGrouping =
  (typeof OverviewGrouping)[keyof typeof OverviewGrouping];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OverviewGrouping = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
} as const;