// src/api/orval/model/getApiOpenVpnClientsOverviewUsersSeriesParams.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { EnumsOverviewGrouping } from './enumsOverviewGrouping';

export type GetApiOpenVpnClientsOverviewUsersSeriesParams = {
From: string;
To: string;
Grouping?: EnumsOverviewGrouping;
VpnServerId?: number;
ExternalId?: string;
};