// src/api/orval/model/overviewUsersSeriesResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { OverviewMetaDto } from "./overviewMetaDto";
import type { OverviewUsersSeriesSummaryDto } from "./overviewUsersSeriesSummaryDto";
import type { OverviewUsersSeriesRowDto } from "./overviewUsersSeriesRowDto";

export interface OverviewUsersSeriesResponse {
  meta?: OverviewMetaDto;
  summary?: OverviewUsersSeriesSummaryDto;
  /** @nullable */
  rows?: OverviewUsersSeriesRowDto[] | null;
}