// src/api/orval/model/overviewUsersSeriesResponseApiResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { OverviewUsersSeriesResponse } from "./overviewUsersSeriesResponse";

export interface OverviewUsersSeriesResponseApiResponse {
  success?: boolean;
  /** @nullable */
  message?: string | null;
  data?: OverviewUsersSeriesResponse;
}